requirements.txt
fastapi
uvicorn
python-socketio
python-engineio
